- major fight with the borrow checker ended  - backtracking added  - updating based on events added
