+++ src/debugger.rs
@@ -1,1 +1,2 @@
 pub mod disassembler;
+pub mod debugger;
