+++ src/main.rs
@@ -1,7 +1,8 @@
-use debugger::disassembler::State;
+use debugger::debugger::State;
 
 pub mod hardware;
 pub mod debugger;
 fn main() {
-    State::start().unwrap();
+    let args: Vec<String> = std::env::args().collect();
+    State::start(args[1].clone()).unwrap();
 }
