+++ src/test_program.txt
@@ -1,0 +1,1 @@
+0123456789
